Fix credit discounts
Stripe doesn't support easilily applying a discount to the total amount
dynamically (yet). So now we're just taking the total amount, the amount
of desired discount, calculating the ratio, and discounting each line
item individually like a bunch of neanderthals.. It works, but is a bit
annoying and we might have to deal with rounding errors better.