Bugfix: project.order always returning the first order
When a Project is part of an Order through OrderItem, and the Stripe
Checkout Session is abandoned, the OrderItems remain. When a new Stripe
Checkout Session is then created, a new Order is created. When the
session is completed, this is the Order we want to see when we call
`project.order` but it would still show the first abandoned Order.